Refactor introduction de sous-modules dans Synchronizer
