+++ app/jobs/synchronizer/communes/api_client_annuaire_administration.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Synchronizer
+  module Communes
+    class ApiClientAnnuaireAdministration
+      include ApiClientOpendatasoftConcern
+
+      def initialize
+        @host = "api-lannuaire.service-public.fr"
+        @dataset_name = "api-lannuaire-administration"
+        @csv_filesize_approximate_in_kb = 120_000
+      end
+    end
+  end
+end
