+++ lib/tasks/objets.rake
@@ -20,7 +20,7 @@ class PalissyStats
     @file = File.open(@path, "wb")
     @total = Objet.count
     log "total: #{@total} objets"
-    (Synchronizer::ObjetBuilder::ALL_FIELDS + ["REFA"]).each { log_field(_1) }
+    (Synchronizer::Objets::Builder::ALL_FIELDS + ["REFA"]).each { log_field(_1) }
     log_photos
     @file.close
   end
